rebase from icg on 0505
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-35496
Reviewed-on: https://android.intel.com:443/363363

+++ camera1/libacc/libacc.cpp
@@ -394,7 +394,7 @@ status_t CameraAcc::handleExit()
  *
  * After we stored the Pointer, we can unblock the calling function acc_map
  **/
-void CameraAcc::notifyPointer(int32_t data, int32_t idx)
+void CameraAcc::notifyPointer(intptr_t data, int32_t idx)
 {
     LOG1("@%s, data=%x, idx=%d", __FUNCTION__, data, idx);
